bot/requirements/requirements.txt
Flask==0.12
python-dotenv==0.6.5
redis==2.10.6
requests==2.18.4
thrift==0.10.0
furl==1.0.1
pymongo==3.5.1
boto3==1.4.7
gunicorn==19.7.1
